Move `layers` into `utils`
