+++ src/otx/algo/detection/necks/yolox_pafpn.py
@@ -11,7 +11,7 @@ from typing import Any
 import torch
 from torch import Tensor, nn
 
-from otx.algo.detection.layers import CSPLayer
+from otx.algo.detection.utils.layer import CSPLayer
 from otx.algo.modules.base_module import BaseModule
 from otx.algo.modules.conv_module import ConvModule
 from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
