+++ src/otx/algo/detection/utils/layer/channel_attention_layer.py
