+++ src/otx/algo/detection/utils/layer/csp_layer.py
@@ -10,11 +10,10 @@ from typing import TYPE_CHECKING
 import torch
 from torch import Tensor, nn
 
+from otx.algo.detection.utils.layer import ChannelAttention
 from otx.algo.modules.base_module import BaseModule
 from otx.algo.modules.conv_module import ConvModule
 from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
-
-from .channel_attention_layer import ChannelAttention
 
 if TYPE_CHECKING:
     from omegaconf import DictConfig
