+++ tests/unit/algo/detection/layers/test_csp_layer.py
@@ -3,8 +3,8 @@
 """Test of CSPLayer."""
 
 import torch
-from otx.algo.detection.layers import ChannelAttention, CSPLayer
-from otx.algo.detection.layers.csp_layer import CSPNeXtBlock, DarknetBottleneck
+from otx.algo.detection.utils.layer import ChannelAttention
+from otx.algo.detection.utils.layer.csp_layer import CSPLayer, CSPNeXtBlock, DarknetBottleneck
 from otx.algo.modules.activation import Swish
 from otx.algo.modules.conv_module import ConvModule
 from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
